Add Interrupt Flag Caching. PPU optimizations
Interrupt flag fetching (through the memory controller, which is really slow) is a major issue. This allows the flags to be Cached on Write.
Also, in the PPU, we avoid memory controller lookups by using OAM and BGRAM lookups directly. This makes it O(1) instead of O(n). We also cache the scanline, saving another memory lookup.